requirements.txt
flask>=3.1
openpyxl>=3.1
nidaqmx>=1.0
pyvisa>=1.13
pyserial>=3.5
